server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pytest==7.4.3
black==22.3.0
freezegun==1.2.2
python-dateutil==2.8.2
Werkzeug==3.0.1